+++ lib/core/facets/hash/setter.rb
@@ -1,7 +1,7 @@
 class Hash
 
-  # Constructs a Proc object from a hash such that the parameter
-  # of the Proc is assigned the hash keys as attributes.
+  # Constructs a Proc that assigns the hash's key-value pairs as
+  # attributes on a given object.
   #
   #   c = Class.new do
   #     attr_accessor :a
@@ -9,20 +9,20 @@ class Hash
   #
   #   h = {:a => 1}
   #   o = c.new
-  #   h.to_proc.call(o)
+  #   h.setter.call(o)
   #   o.a  #=> 1
   #
-  # If +response+ is set to +true+, then assignment will only occur
-  # if receiver responds_to? the writer method.
+  # If +safe+ is set to +true+, then assignment will only occur
+  # if the receiver responds_to? the writer method.
   #
   # CREDIT: Trans
 
-  def to_proc(response=false)
-    if response
+  def setter(safe=false)
+    if safe
       lambda do |o|
         self.each do |k,v|
           ke = "#{k}="
-          o.__send__(ke, v) if respond_to?(ke)
+          o.__send__(ke, v) if o.respond_to?(ke)
         end
       end
     else
@@ -36,4 +36,3 @@ class Hash
   end
 
 end
-